distributed-systems/assignment-2/cassandra-setup/requirements.txt
-i https://pypi.org/simple
cassandra-driver==3.29.2
certifi==2025.10.5; python_version >= '3.7'
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.3.0; python_version >= '3.10'
geomet==0.2.1.post1; python_version >= '2.7' and python_version != '3.3' and python_version < '4'
idna==3.11; python_version >= '3.8'
requests==2.32.5; python_version >= '3.9'
structlog==25.4.0; python_version >= '3.8'
urllib3==2.5.0; python_version >= '3.9'
